- Added referential integrity checks to DB - Prevent user from deleting running program and informing via alert - Fixed tests

+++ picure/API/program.py
@@ -14,7 +14,9 @@
 #      You should have received a copy of the GNU General Public License
 #      along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import json
-from flask import Blueprint, request, Response
+import sqlite3
+
+from flask import Blueprint, request, Response, abort
 from picure.Backend.Program import controler
 from picure.Backend.DB.db_handler import get_db
 
@@ -23,8 +25,10 @@ program = Blueprint("program", __name__, template_folder="templates")
 
 @program.route("/api/program", methods=["GET"])
 def get_programs():
-    progs = [{"id": t.program_id, "name": t.name, "steps": len(t.get_steps())} for t in
-             controler.get_list_of_programs()]
+    progs = [
+        {"id": t.program_id, "name": t.name, "steps": len(t.get_steps())}
+        for t in controler.get_list_of_programs()
+    ]
     return json.dumps(progs)
 
 
@@ -33,7 +37,9 @@ def new_program():
     form_data = request.form.to_dict()
     if "ProgramName" in form_data:
         db = get_db()
-        db.cursor().execute("INSERT INTO program (name) VALUES (:ProgramName)", form_data )
+        db.cursor().execute(
+            "INSERT INTO program (name) VALUES (:ProgramName)", form_data
+        )
         db.commit()
         db.cursor().close()
     return Response(response="No Content", status=204)
@@ -42,7 +48,16 @@ def new_program():
 @program.route("/api/program/<string:prog_id>", methods=["DELETE"])
 def delete_program(prog_id):
     db = get_db()
-    db.cursor().execute('DELETE from program where id = ?', (prog_id,) )
+    if (
+        db.cursor()
+        .execute(
+            "SELECT max(enabled) as enabled from program_run where program_id = ?", (prog_id,)
+        )
+        .fetchone()['enabled']
+        == 1
+    ):
+        abort(500, "Cant delete currently running program")
+    db.cursor().execute("DELETE from program where id = ?", (prog_id,))
     db.commit()
     db.cursor().close()
     return "Ok"
